Reduce memory usage of eachMapping w/ loop
Avoid creating arrays of intermediate values to pass to `aCallback`, by
refactoring the .map + .forEach combination to a single loop that that
creates the options object and immediately passes it to aCallback.

This reduces memory usage, because engine no longer has to hold all the
options objects in memory, each of the created objects is free to be gcd
once the loop continues to next item.

When the looped-over array `mappings` is very large (e.g. >40k items),
the operation may even crash with out of memory exception without this
patch.